requirements.txt
# Core dependencies
python-dotenv==1.0.0
toml==0.10.2

# Binance
python-binance==1.0.19
websockets==12.0
binance-connector==3.5.0

# Telegram
python-telegram-bot==20.7

# Technical Analysis
TA-Lib==0.4.28
pandas==2.1.4
numpy==1.26.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Logging
python-json-logger==2.0.7

# Utilities
pytz==2023.3